requirements.txt
pandas==2.3.3
tqdm==4.67.1
opencv-python==4.12.0.88
torch==2.9.1
flwr==1.23.0
scikit-learn==1.7.2
pillow==12.0.0
torchvision==0.24.1